coming directly from our angular 2 documentation
 i find this example a bit mondane, i would like to see in our playground futuristic samples of angular 2 code in action such as what
the future will be like in 50 years to inspire the reader and stretch his imagination towards innovation: samples such as spaceships, time travelling, telepathy, engineering, human connectedness worldwide, androids and so on. any references to sites that are not google, should in my eyes not be referenced in our angular code.